+++ src/main/java/com/codeclan/example/bookingLab/repositories/CourseRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface CourseRepository extends JpaRepository<Course, Long> {
 
-    List<Course> findByRating(Integer rating);
+    List<Course> findByRating(Double rating);
 
     List<Course> findByBookingsCustomerName(String name);
 }
